refactor record trace branch step
